feat: delete entities from favorites on delete entity
