+++ src/modules/db/entities/album/album.entity.ts
@@ -1,12 +1,20 @@
-import { Injectable } from '@nestjs/common';
-import { Album } from 'src/models';
+import { Inject, Injectable, forwardRef } from '@nestjs/common';
+import { Album, Track } from 'src/models';
 import { CreateAlbumDto } from 'src/modules/album/dto/create-album.dto';
+import { FavoritesEntity } from '../favorites/favorites.entity';
+import { TrackEntity } from '../track/track.entity';
 
 @Injectable()
 export class AlbumEntity {
   private album: Album[] = [];
 
-  constructor() {}
+  constructor(
+    @Inject(forwardRef(() => FavoritesEntity))
+    private readonly favoritesEntity: FavoritesEntity,
+
+    @Inject(forwardRef(() => TrackEntity))
+    private readonly trackEntity: TrackEntity,
+  ) {}
 
   create(createAlbumDto: CreateAlbumDto): Album {
     const album: Album = new Album(createAlbumDto);
@@ -45,6 +53,18 @@ export class AlbumEntity {
 
     if (albumIndex !== -1) {
       this.album.splice(albumIndex, 1);
+
+      const albumTracks: Track[] = this.trackEntity.findAllByAlbumId(id);
+
+      // TODO: refactor using Prisma
+      albumTracks.forEach((track) => {
+        this.trackEntity.update(track.id, {
+          ...track,
+          albumId: null,
+        });
+      });
+
+      this.favoritesEntity.deleteAlbum(id);
 
       return;
     }
